Implemented getNSimilarUsers
