+++ src/entity/UserSimilarity.java
@@ -1,20 +1,20 @@
-package entity;
-
-public class UserSimilarity {
-    private final String username;
-    private final double score;
-
-    public UserSimilarity(String username, double score) {
-        this.username = username;
-        this.score = score;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public double getScore() {
-        return score;
-    }
-}
-
+//package entity;
+//
+//private static class UserSimilarity {
+//    private final String username;
+//    private final double score;
+//
+//    public UserSimilarity(String username, double score) {
+//        this.username = username;
+//        this.score = score;
+//    }
+//
+//    public String getUsername() {
+//        return username;
+//    }
+//
+//    public double getScore() {
+//        return score;
+//    }
+//}
+//
